python_template/requirements.txt
# Python dependencies for the Phytovenomics ML Platform with ESM-2 transformer implementation

# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Deep learning
torch>=1.10.0
transformers>=4.18.0,<4.30.0  # For ESM-2 model, version limit for compatibility
datasets>=2.0.0
accelerate>=0.10.0  # For mixed-precision training
fair-esm>=2.0.0  # Facebook Research ESM models
huggingface-hub>=0.12.0  # For model downloading

# Bioinformatics
biopython>=1.79  # For sequence analysis
biotite>=0.33.0  # For structural analysis

# Natural Language Processing
regex>=2022.1.18  # For advanced pattern matching
nltk>=3.6.0

# Utilities
pyyaml>=6.0  # For configuration files
tqdm>=4.62.0  # For progress bars
joblib>=1.1.0  # For parallel processing
tensorboard>=2.8.0  # For training visualization
pytest>=6.2.5  # For testing

# Antibody-specific libraries
Bio-AbWrapper>=0.1.0  # For antibody numbering schemes (Kabat, Chothia, etc.)
IGoR>=1.4.0  # For immunoglobulin sequence analysis

# Optimization
optuna>=2.10.0  # For hyperparameter optimization
ray[tune]>=1.13.0  # For distributed hyperparameter tuning

# Visualization
plotly>=5.5.0
networkx>=2.6.3  # For network visualization

# Development tools
black>=22.1.0  # Code formatting
isort>=5.10.1  # Import sorting
flake8>=4.0.1  # Linting